Authentication using Sanctum

+++ routes/api.php
@@ -10,19 +10,22 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
 });
 
 Route::get('/ping', function (Request $request) {
-    return [
-        'pong' => true
-    ];
+    return ['pong' => true];
 });
+Route::get('/unauthenticated', function () {
+    return ['error' => 'User not logged!'];
+})->name('login');
 
 Route::post('/user', [AuthController::class, 'create']);
+Route::middleware('auth:sanctum')->get('/auth/logout', [AuthController::class, 'logout']);
+Route::post('/auth', [AuthController::class, 'login']);
 
-Route::post('/todo', [ApiController::class, 'store']);
+Route::middleware('auth:sanctum')->post('/todo', [ApiController::class, 'store']);
 
 Route::get('/todos', [ApiController::class, 'index']);
 
 Route::get('/todo/{id}', [ApiController::class, 'show']);
 
-Route::put('/todo/{id}', [ApiController::class, 'update']);
+Route::middleware('auth:sanctum')->put('/todo/{id}', [ApiController::class, 'update']);
 
-Route::delete('/todo/{id}', [ApiController::class, 'delete']);
+Route::middleware('auth:sanctum')->delete('/todo/{id}', [ApiController::class, 'delete']);
